fix incremental_learning bug - Add docs and code comment - fix bugs: epoch always be 1, inference result not saved
