+++ examples/incremental_learning/helmet_detection/training/inference.py
@@ -25,10 +25,11 @@ from sedna.core.incremental_learning import IncrementalLearning
 from interface import Estimator
 
 
-he_saved_url = Context.get_parameters("HE_SAVED_URL")
+he_saved_url = Context.get_parameters("HE_SAVED_URL", '/tmp')
+rsl_saved_url = Context.get_parameters("RESULT_SAVED_URL", '/tmp')
 class_names = ['person', 'helmet', 'helmet_on', 'helmet_off']
 
-FileOps.clean_folder([he_saved_url], clean=False)
+FileOps.clean_folder([he_saved_url, rsl_saved_url], clean=False)
 
 
 def draw_boxes(img, labels, scores, bboxes, class_names, colors):
@@ -72,12 +73,13 @@ def output_deal(is_hard_example, infer_result, nframe, img_rgb):
     img_rgb = np.array(img_rgb)
     img_rgb = cv2.cvtColor(img_rgb, cv2.COLOR_RGB2BGR)
     colors = 'yellow,blue,green,red'
-    if not is_hard_example:
-        return
+
     lables, scores, bbox_list_pred = infer_result
     img = draw_boxes(img_rgb, lables, scores, bbox_list_pred, class_names,
                      colors)
-    cv2.imwrite(f"{he_saved_url}/{nframe}.jpeg", img)
+    if is_hard_example:
+        cv2.imwrite(f"{he_saved_url}/{nframe}.jpeg", img)
+    cv2.imwrite(f"{rsl_saved_url}/{nframe}.jpeg", img)
 
 
 def mkdir(path):
