+++ lib/sedna/core/incremental_learning/incremental_learning.py
@@ -28,9 +28,13 @@ class IncrementalLearning(JobBase):
     Incremental learning
     """
 
-    def __init__(self, estimator, config=None):
-        super(IncrementalLearning, self).__init__(
-            estimator=estimator, config=config)
+    def __init__(self, estimator):
+        """
+        Initial a IncrementalLearning job
+        :param estimator: Customize estimator
+        """
+
+        super(IncrementalLearning, self).__init__(estimator=estimator)
 
         self.model_urls = self.get_parameters(
             "MODEL_URLS")  # use in evaluation
@@ -42,6 +46,15 @@ class IncrementalLearning(JobBase):
               valid_data=None,
               post_process=None,
               **kwargs):
+        """
+        Training task for IncrementalLearning
+        :param train_data: datasource use for train
+        :param valid_data: datasource use for evaluation
+        :param post_process: post process
+        :param kwargs: params for training of customize estimator
+        :return: estimator
+        """
+
         callback_func = None
         if post_process is not None:
             callback_func = ClassFactory.get_cls(
@@ -58,6 +71,14 @@ class IncrementalLearning(JobBase):
             self.estimator) if callback_func else self.estimator
 
     def inference(self, data=None, post_process=None, **kwargs):
+        """
+        Inference task for IncrementalLearning
+        :param data: inference sample
+        :param post_process: post process
+        :param kwargs: params for inference of customize estimator
+        :return: inference result, result after post_process, if is hard sample
+        """
+
         if not self.estimator.has_load:
             self.estimator.load(self.model_path)
 
@@ -81,6 +102,14 @@ class IncrementalLearning(JobBase):
         return infer_res, res, is_hard_example
 
     def evaluate(self, data, post_process=None, **kwargs):
+        """
+        Evaluate task for IncrementalLearning
+        :param data: datasource use for evaluation
+        :param post_process: post process
+        :param kwargs: params for evaluate of customize estimator
+        :return: evaluate metrics
+        """
+
         callback_func = None
         if callable(post_process):
             callback_func = post_process
